refactored a lot of small things, added datagen data for the reacting recipes, added suppression for a lot of warnings, change the gunship font to use the new modid

+++ src/main/java/com/Ultra_Nerd/CodeLyokoLegacy/util/AdditionalStateProperties.java
@@ -25,7 +25,7 @@ public record AdditionalStateProperties() {
                 final Set<Float> set = Sets.newHashSet();
 
                 for(float i = min; i <= max; ++i) {
-                    set.add(Float.valueOf(i));
+                    set.add(i);
                 }
 
                 this.values = ImmutableSet.copyOf(set);
@@ -95,7 +95,7 @@ public record AdditionalStateProperties() {
                 final Set<Double> set = Sets.newHashSet();
 
                 for(double i = min; i <= max; ++i) {
-                    set.add(Double.valueOf(i));
+                    set.add(i);
                 }
 
                 this.values = ImmutableSet.copyOf(set);
